Move HcloudLogger interface to Base module
