--- src/lib/interfaces/global/Logger.ts
@@ -1,6 +1,0 @@
-export interface HcloudLogger {
-    info(msg: string): void;
-    debug(msg: string): void;
-    warn(msg: string): void;
-    error(msg: string): void;
-}
