+++ src/lib/interfaces/global/index.ts
@@ -1,7 +1,6 @@
 export * from "./ErrorMessage";
 export * from "./Header";
 export * from "./HttpMethod";
-export * from "./Logger";
 export * from "./Products";
 export * from "./SearchFilters";
 export * from "./Space";
